+++ gsa/src/web/pages/scanconfigs/editnvtdetailsdialog.js
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-import React, {useEffect, useState} from 'react';
+import React, {useEffect, useReducer, useState} from 'react';
 
 import _ from 'gmp/locale';
 
@@ -74,6 +74,32 @@ const createPrefValues = (preferences = []) => {
 const convertTimeout = value =>
   !isDefined(value) || value.trim().length === 0 ? undefined : value;
 
+const reducer = (state, action) => {
+  switch (action.type) {
+    case 'setValue':
+      const {newState} = action;
+      const {name, value} = newState;
+
+      const prefAttributes = state[name];
+
+      // preference has other attributes like id, type, etc. those must be kept so its not as simple as [name]: value
+      const updatedState = {
+        ...state,
+        [name]: {
+          ...prefAttributes,
+          value,
+        },
+      };
+
+      return updatedState;
+    case 'setAll':
+      const {formValues} = action;
+      return formValues;
+    default:
+      return state;
+  }
+};
+
 const EditNvtDetailsDialog = ({
   configId,
   configName,
@@ -96,7 +122,8 @@ const EditNvtDetailsDialog = ({
 }) => {
   timeout = convertTimeout(timeout);
 
-  const [preferenceValues, setPreferenceValues] = useState(
+  const [preferenceValues, dispatch] = useReducer(
+    reducer,
     createPrefValues(preferences),
   );
 
@@ -106,7 +133,10 @@ const EditNvtDetailsDialog = ({
   );
 
   useEffect(() => {
-    setPreferenceValues(createPrefValues(preferences));
+    dispatch({
+      type: 'setAll',
+      formValues: createPrefValues(preferences),
+    });
   }, [preferences]);
 
   useEffect(() => {
@@ -268,9 +298,7 @@ const EditNvtDetailsDialog = ({
                       key={pref.name}
                       preference={pref}
                       value={prefValue}
-                      onChange={value => {
-                        preferenceValues[pref.name].value = value.value;
-                      }}
+                      onChange={dispatch}
                     />
                   );
                 })}